Update Hugging Face model and URL for summarization
What the title says lol. I just changed the URL to send a request for summarization to the correct one (because the technology is changing so fast, the documentation is a bit of a mess, and it gets hard to tell what is current and what is deprecated. I also changed the AI to one that is actually supported by the Hugging Face Inference API (Free).